# backend/docs/swagger.yaml
definitions:
  domain.User:
    properties:
      birth_date:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
    required:
    - birth_date
    - email
    type: object
info:
  contact: {}
paths:
  /users:
    get:
      consumes:
      - application/json
      description: Get all users
      produces:
      - application/json
      responses:
        "200":
          description: Users found
          schema:
            items:
              $ref: '#/definitions/domain.User'
            type: array
        "500":
          description: Internal server error
          schema: {}
      summary: Get all users
    post:
      consumes:
      - application/json
      description: Creates a new user
      parameters:
      - description: User to create
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.User'
      produces:
      - application/json
      responses:
        "201":
          description: User created
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Invalid request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Creates a new user
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: User deleted
        "400":
          description: Invalid ID
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Delete a user
    get:
      consumes:
      - application/json
      description: Get a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User found
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Invalid ID
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Get a user by ID
    put:
      consumes:
      - application/json
      description: Update a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.User'
      produces:
      - application/json
      responses:
        "200":
          description: User updated
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Invalid ID
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Update a user
swagger: "2.0"